+++ src/components/molecules/FormEditor/FormEditor.tsx
@@ -3,7 +3,7 @@ import {useDebounce} from 'react-use';
 
 // @ts-ignore
 import {Theme as AntDTheme} from '@rjsf/antd';
-import {ObjectFieldTemplateProps, withTheme} from '@rjsf/core';
+import {withTheme} from '@rjsf/core';
 
 import fs from 'fs';
 import log from 'loglevel';
@@ -28,6 +28,7 @@ import {CHANGES_BY_FORM_EDITOR, trackEvent} from '@utils/telemetry';
 import {parseYamlDocument} from '@utils/yaml';
 
 import * as S from './FormEditor.styled';
+import FormObjectFieldTemplate from './FormObjectFieldTemplate';
 import {getCustomFormFields, getCustomFormWidgets} from './FormWidgets';
 
 const Form = withTheme(AntDTheme);
@@ -182,41 +183,13 @@ const FormEditor: React.FC<IProps> = props => {
     }
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-shadow
-  const ObjectFieldTemplate = (props: ObjectFieldTemplateProps) => {
-    const {title, properties, uiSchema} = props;
-    const [isExpanded, toggleExpand] = useState<boolean>(true);
-    const opacity = (10 - (uiSchema?.level ?? 0)) / 10;
-
-    return (
-      <S.FieldContainer>
-        <S.TitleWrapper onClick={() => toggleExpand(prev => !prev)} opacityStep={opacity || 1}>
-          {isExpanded ? <S.ArrowIconExpanded /> : <S.ArrowIconClosed />}
-
-          {title ? (
-            <S.TitleText isBold={uiSchema.level === 0}>{title}</S.TitleText>
-          ) : (
-            <S.ElementText>element</S.ElementText>
-          )}
-        </S.TitleWrapper>
-        {isExpanded && (
-          <>
-            {properties.map((element: any) => (
-              <S.PropertyContainer key={element.content.key}>{element.content}</S.PropertyContainer>
-            ))}
-          </>
-        )}
-      </S.FieldContainer>
-    );
-  };
-
   return (
     <S.FormContainer>
       <Form
         schema={schema}
         uiSchema={formUiSchema}
         formData={formData}
-        ObjectFieldTemplate={ObjectFieldTemplate}
+        ObjectFieldTemplate={FormObjectFieldTemplate}
         onChange={onFormUpdate}
         widgets={getCustomFormWidgets()}
         fields={getCustomFormFields()}
